Init PIN if missing. Moved mdk func to makeDESkeys
Consolidated two scripts into one: makeDESkeys and mdk
are now colocated in makeDESkeys. Install process now
links that to mdk. When invoked as 'mdk', the script
produces /n/ keys; when invoked as makeDESkeys, the
script produces a single key on the clipboard.

All (both) scripts now reside in bin.

If password (PIN) file doesn't exist, it prompts to create
it for you.