Add read encrypted save file feat + pbcopy support
In addition to creating a new key (or series of keys), you can now read
from an encrypted or clear-text list of keys using a maintained index
file to keep track of your place in the file.

readDESkeys does the same thing as makeDESkeys, but using this
file-based concept instead of generating new keys.

Also supports Mac OS X clipboard (pbcopy) utility as well as Linux xclip
to get the result to the clipboard.